pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aa-moonmining"
dynamic = ["version"]
description = "Alliance Auth app for tracking moon extractions and scouting new moons."
readme = "README.md"
license = "MIT"
requires-python = ">=3.8"
authors = [
    { name = "Erik Kalkoken", email = "kaloken87@gmail.com" },
]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 4.0",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]
dependencies = [
    "allianceauth-app-utils>=1.14.2",
    "allianceauth>=3",
    "django-bootstrap-form",
    "django-datatables-view>=1.20",
    "django-eveuniverse>=0.18",
    "django-navhelper",
    "PyYAML",
]

[project.urls]
Homepage = "https://gitlab.com/ErikKalkoken/aa-moonmining"
Source = "https://gitlab.com/ErikKalkoken/aa-moonmining"
Changelog = "https://gitlab.com/ErikKalkoken/aa-moonmining/-/blob/master/CHANGELOG.md"
Tracker = "https://gitlab.com/ErikKalkoken/aa-moonmining/-/issues"


[tool.hatch.version]
path = "moonmining/__init__.py"

[tool.hatch.build]
include = [
    "/moonmining",
]

[tool.isort]
profile = "black"
multi_line_output = 3
sections = [
    "FUTURE",
    "STDLIB",
    "THIRDPARTY",
    "DJANGO",
    "ALLIANCEAUTH",
    "FIRSTPARTY",
    "LOCALFOLDER"
]
known_allianceauth = ["allianceauth", "app_utils"]
known_django = ["django", "django_webtest", "esi", "eveuniverse"]

[tool.pylint.'MASTER']
ignore-patterns = ["test_.*.py,__init__.py", "generate_.*.py"]

[tool.pylint.'BASIC']
good-names = ["i", "j", "k", "x", "f", "ex"]

[tool.pylint.'FORMAT']
max-line-length = 100

[tool.pylint.'MESSAGES CONTROL']
disable = ["R", "C"]
